1. camp 생성 수정 삭제 추가 댓글 생성 버그 수정 / heart 생성중

+++ project/src/main/java/com/campers/camfp/service/camp/CampServiceImpl.java
@@ -234,7 +234,7 @@ public class CampServiceImpl implements CampService {
 			break;
 		}
 
-		return (Long) result;
+		return result;
 
 	}
